penjualan

+++ database/migrations/2022_11_12_131211_create_detail_penjualan_temps_table.php
@@ -18,7 +18,8 @@ class CreateDetailPenjualanTempsTable extends Migration
             $table->foreignId('produk_id')->constrained('produks');
             $table->float('qty');
             $table->string('ket', 50);
-            $table->integer('disc');
+            // $table->integer('disc');
+            $table->decimal('disc', 15, 1);
             $table->decimal('jumlah', 15, 1);
             $table->timestamps();
         });
